Use execvp instead of execv
execvp searches a user's $PATH, allowing for the specification of
relative paths, which this commit uses by default. This is particularly
convenient for users and administrators who wish to set up portable
installations or have systems with mercurial filesystem layouts.

In introducing this change, the path for `asl` is no longer hard-coded
to `/usr/bin/asl`; instead a macro `ASEXE` can be provided. Since its
invocations also use execvp, it can be anywhere on $PATH, and thus
`ASEXE` defaults to the empty string.