Remove redundant dubdabx
The lcc1802epilo{DH,NG,NO}.inc files have admitted that the "double
dabble" algorithm better fit into nstdlib.inc, where it was introduced
in "20-08-01" according to its changelog. The redundancy results in an
error in asl 1.42 Beta (current as of this writing), so it may as well
be removed now.